+++ demo/site-pages/src/documents/pages/blocks/StageBlock.tsx
@@ -1,8 +1,0 @@
-import { ListBlock, type PropsWithData } from "@comet/site-nextjs";
-import { type StageBlockData } from "@src/blocks.generated";
-
-import { BasicStageBlock } from "./BasicStageBlock";
-
-export const StageBlock = ({ data }: PropsWithData<StageBlockData>) => {
-    return <ListBlock data={data} block={(block) => <BasicStageBlock data={block} />} />;
-};