+++ packages/api/cms-api/src/dam/files/files.service.ts
@@ -4,11 +4,11 @@ import { forwardRef, Inject, Injectable, Logger } from "@nestjs/common";
 import { createHmac } from "crypto";
 import exifr from "exifr";
 import { createReadStream } from "fs";
-import getColors from "get-image-colors";
 import * as hasha from "hasha";
 import { basename, extname, parse } from "path";
 import probe from "probe-image-size";
 import * as rimraf from "rimraf";
+import sharp from "sharp";
 
 import { BlobStorageBackendService } from "../../blob-storage/backends/blob-storage-backend.service";
 import { createHashedPath } from "../../blob-storage/utils/create-hashed-path.util";
@@ -24,6 +24,7 @@ import { CometImageResolutionException } from "../common/errors/image-resolution
 import { DamConfig } from "../dam.config";
 import { DAM_CONFIG } from "../dam.constants";
 import { ImageCropAreaInput } from "../images/dto/image-crop-area.input";
+import { rgbToHex } from "../images/images.util";
 import { DamScopeInterface } from "../types";
 import { DamMediaAlternative } from "./dam-media-alternatives/entities/dam-media-alternative.entity";
 import { DamFileListPositionArgs, FileArgsInterface } from "./dto/file.args";
@@ -545,7 +546,7 @@ export class FilesService {
     async calculateDominantColor(contentHash: string): Promise<string | undefined> {
         const path = this.imgproxyService
             .builder()
-            .resize(ResizingType.AUTO, 128)
+            .resize(ResizingType.AUTO, 1)
             .format(Extension.JPG)
             .generateUrl(
                 `${this.blobStorageBackendService.getBackendFilePathPrefix()}${this.config.filesDirectory}/${createHashedPath(contentHash)}`,
@@ -557,13 +558,12 @@ export class FilesService {
             const arrayBuffer = await imageResponse.arrayBuffer();
             const imageBuffer = Buffer.from(arrayBuffer);
 
-            // @TODO: make pull request to fix overloads https://github.com/DefinitelyTyped/DefinitelyTyped/blob/master/types/get-image-colors/index.d.ts#L15
-            // eslint-disable-next-line @typescript-eslint/no-explicit-any
-            const getColorOptions: any = { type: "image/jpg", count: 1 };
+            const data = await sharp(imageBuffer).removeAlpha().raw().toBuffer();
 
-            const colors = (await getColors(imageBuffer, getColorOptions)).map((color) => color.hex());
+            // since we scale the image to 1px with imgproxy, data only contains the colors for this one pixel
+            const [r, g, b] = data;
 
-            return colors.length > 0 ? colors[0] : undefined;
+            return rgbToHex(r, g, b);
         } catch (e) {
             // When imageproxy is not available it is ok.
             console.error(e);