+++ packages/eslint-config/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @comet/eslint-config
 
+## 8.7.1
+
+### Patch Changes
+
+- @comet/eslint-plugin@8.7.1
+
+## 8.7.0
+
+### Patch Changes
+
+- @comet/eslint-plugin@8.7.0
+
 ## 8.6.0
 
 ### Patch Changes